+++ src/main/java/dev/fabriciosilva/webservice/domain/notafiscal/dto/NotaFiscalForm.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public class NotaFiscalForm {
 
-    private Integer numero;
+    private Long numero;
     private ClienteInfo cliente;
     private LocalDate data;
     private List<ItemForm> itens;
@@ -16,11 +16,11 @@ public class NotaFiscalForm {
     public NotaFiscalForm() {
     }
 
-    public Integer getNumero() {
+    public Long getNumero() {
         return numero;
     }
 
-    public void setNumero(Integer numero) {
+    public void setNumero(Long numero) {
         this.numero = numero;
     }
